キズ一覧をコンポーネント化 #203
